backend/source/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
deepdiff==7.0.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.6
motor==3.4.0
numpy==1.26.4
ordered-set==4.1.0
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.24.0.post1
